backend/requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# NLP
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.0/en_core_web_sm-3.7.0-py3-none-any.whl

# Validation & Serialization
pydantic==2.5.3
pydantic-settings==2.1.0

# Caching
redis==5.0.1
hiredis==2.3.2

# HTTP Client
httpx==0.26.0

# Image Processing
pillow==10.2.0

# Data Processing
datasets==2.16.1
pandas==2.1.4

# Utilities
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
tqdm==4.66.1
mwclient==0.10.1

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.1
ruff==0.1.14
mypy==1.8.0
